IN-198 fix: remove console error for postStoreProvider, create PostHeaderPopover, and move logic inside usePostContent
