+++ src/features/posts/ui/PostModal/Content/index.ts
@@ -5,3 +5,4 @@ export * from './LikesGroup'
 export * from './PostContent'
 export * from './PostInfo'
 export * from './SocialGroup'
+export * from './usePostContent'
