+++ src/features/posts/ui/PostModal/Header/index.ts
@@ -1,1 +1,2 @@
+export * from './PostHeaderPopover'
 export * from './PostInfoHeader'
